@@ -1,111 +1,80 @@
 {
-    "name": "node-backend-starter",
-    "version": "0.0.1",
-    "private": true,
-    "author": "viettd",
+    "name": "src",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "directories": {
+        "doc": "docs"
+    },
     "scripts": {
-        "start": "pm2-runtime start /app/ecosystem.config.js",
-        "dev": "NODE_ENV=dev DEBUG=api:* DOTENV_FILE=.env pm2-runtime start ./ecosystem.dev.config.js",
-        "seed_db": "NODE_ENV=test DOTENV_FILE=.env.test ts-node ./src/seedDB.ts",
-        "test": "NODE_ENV=test DOTENV_FILE=.env.test jest --maxWorkers=4 --forceExit",
-        "test_ci": "NODE_ENV=test_ci DOTENV_FILE=.env.ci jest --maxWorkers=4 --forceExit",
-        "watch": "concurrently \"gulp init\" \"ttsc -w -p ttsconfig.dev.json\" \"graphql-codegen --config codegen-watch.yml\" \"yarn gulp\"",
-        "codegen": "graphql-codegen --config codegen.yml",
-        "build": "gulp init && yarn codegen && ttsc -p ttsconfig.json"
+        "build": "ttsc",
+        "watch": "ttsc -w",
+        "dev_test": "node -r dotenv/config",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "devDependencies": {
+        "@types/aws4": "^1.11.6",
+        "@types/bcryptjs": "^2.4.6",
+        "@types/chance": "^1.1.6",
+        "@types/cors": "^2.8.17",
+        "@types/express": "^4.17.21",
+        "@types/formidable": "^3.4.5",
+        "@types/ioredis": "^5.0.0",
+        "@types/jsonwebtoken": "^9.0.6",
+        "@types/lodash": "^4.14.202",
+        "@types/mime-types": "^2.1.4",
+        "@types/node": "^20.11.24",
+        "@types/rimraf": "^4.0.5",
+        "@types/slug": "^5.0.7",
+        "@types/ua-parser-js": "^0.7.39",
+        "@types/umzug": "^2.3.7",
+        "@types/uuid": "^9.0.8",
+        "@types/validator": "^13.11.9",
+        "ts-transformer-imports": "^0.4.3",
+        "ttypescript": "^1.5.15",
+        "typescript": "4.9.5"
     },
     "dependencies": {
-        "@hapi/joi": "15.1.1",
-        "apollo-server-cache-redis": "^3.3.1",
-        "apollo-server-express": "^3.6.7",
-        "awilix": "^7.0.2",
-        "axios": "0.27.2",
+        "@aternus/csv-to-xlsx": "^2.1.5",
+        "@aws-sdk/client-sqs": "^3.775.0",
+        "@bull-board/express": "5.20.5",
+        "aws4": "^1.12.0",
+        "axios": "^1.8.4",
         "bcryptjs": "^2.4.3",
-        "bull": "4.8.2",
-        "compression": "1.7.4",
-        "cors": "2.8.5",
-        "dataloader": "^2.1.0",
-        "dotenv": "^16.0.0",
-        "escape-string-regexp": "5.0.0",
-        "express": "4.18.0",
-        "extend": "^3.0.2",
-        "graphql": "16.4.0",
-        "graphql-depth-limit": "1.1.0",
-        "graphql-format-error-context-extension": "^0.2.15",
-        "graphql-tools": "8.2.8",
-        "helmet": "5.0.2",
-        "i18next": "^21.6.16",
-        "ioredis": "5.0.4",
-        "json-stringify-safe": "5.0.1",
-        "jsonwebtoken": "8.5.1",
-        "lodash": "4.17.21",
-        "moleculer": "0.14.20",
-        "moleculer-jaeger": "0.2.3",
-        "moment-timezone": "0.5.34",
-        "mongo-uri-builder": "^4.0.0",
-        "mongodb-uri": "^0.9.7",
-        "mongoose": "^6.3.1",
-        "morgan": "1.10.0",
-        "nats": "^2.6.1",
-        "pino": "^7.11.0",
-        "randomatic": "3.1.1",
-        "safe-regex": "2.1.1"
-    },
-    "devDependencies": {
-        "@graphql-codegen/cli": "^2.6.2",
-        "@graphql-codegen/typescript": "2.4.8",
-        "@graphql-codegen/typescript-resolvers": "^2.6.1",
-        "@types/bcryptjs": "^2.4.2",
-        "@types/bull": "^3.15.8",
-        "@types/compression": "^1.7.2",
-        "@types/cors": "^2.8.12",
-        "@types/hapi__joi": "^17.1.8",
-        "@types/helmet": "^0.0.48",
-        "@types/ioredis": "^4.28.10",
-        "@types/jest": "^27.4.1",
-        "@types/json-stringify-safe": "^5.0.0",
-        "@types/jsonwebtoken": "^8.5.8",
-        "@types/lodash": "^4.14.182",
-        "@types/mongodb-uri": "^0.9.1",
-        "@types/morgan": "^1.9.3",
-        "@types/rimraf": "^3.0.2",
-        "@types/safe-regex": "^1.1.3",
-        "@types/uuid": "^8.3.4",
-        "@typescript-eslint/eslint-plugin": "^5.21.0",
-        "@typescript-eslint/parser": "^5.21.0",
-        "concurrently": "^7.1.0",
-        "del": "^6.0.0",
-        "eslint": "^8.14.0",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-prettier": "^4.0.0",
-        "faker": "^6.6.6",
-        "gulp": "^4.0.2",
-        "husky": "^7.0.4",
-        "i18next-scanner": "^3.1.0",
-        "i18next-scanner-typescript": "^1.0.6",
-        "jest": "^28.0.2",
-        "jest-environment-node": "^28.0.2",
-        "lint-staged": "^12.4.1",
-        "mongo-seeding": "^3.7.1",
-        "mongodb-memory-server": "^8.5.1",
-        "nock": "^13.2.4",
-        "nodemon": "^2.0.15",
-        "npm-check-updates": "^12.5.10",
-        "prettier": "^2.6.2",
-        "pretty-quick": "^3.1.3",
-        "ts-jest": "^27.1.4",
-        "ts-node": "^10.7.0",
-        "ts-transformer-imports": "^0.4.3",
-        "ttypescript": "^1.5.13",
-        "typescript": "^4.6.4"
-    },
-    "lint-staged": {
-        "*.ts": [
-            "eslint --fix"
-        ]
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "yarn pretty-quick --staged && lint-staged"
-        }
+        "bullmq": "^5.41.7",
+        "chance": "^1.1.11",
+        "cors": "^2.8.5",
+        "croner": "^8.0.2",
+        "crypto": "^1.0.1",
+        "dotenv": "^16.4.5",
+        "express": "^5.1.0",
+        "express-basic-auth": "^1.2.1",
+        "firebase-admin": "^13.1.0",
+        "formidable": "^3.5.1",
+        "ioredis": "^5.3.2",
+        "joi": "17.4.2",
+        "jquery-csv": "^1.0.21",
+        "jsonwebtoken": "^9.0.2",
+        "lodash": "^4.17.21",
+        "mime-types": "^2.1.35",
+        "moment": "^2.30.1",
+        "moment-timezone": "^0.5.45",
+        "nanoid": "3",
+        "node-cache": "^5.1.2",
+        "pg": "^8.11.3",
+        "pg-connection-string": "^2.6.2",
+        "pg-hstore": "^2.3.4",
+        "reflect-metadata": "^0.2.1",
+        "rimraf": "^6.0.1",
+        "sequelize": "6.37.1",
+        "sequelize-typescript": "^2.1.6",
+        "slug": "^8.2.3",
+        "strong-error-handler": "^5.0.8",
+        "ua-parser-js": "^1.0.37",
+        "umzug": "^3.7.0",
+        "uuid": "^9.0.1"
     }
 }
